Trait quantum_world_state::QWSElement

pub trait QWSElement: Debug {
    fn meta_data_keys_iter(
        &self
    ) -> Option<Box<dyn Iterator<Item = QWSMetaData>>>;

    fn meta_data_values_for_key_iter(
        &self, 
        key: &QWSMetaData
    ) -> Option<Box<dyn Iterator<Item = QWSMetaData>>>;

    fn as_any(&self) -> &dyn Any;
}

An element in the QuantumWorldState must implement this trait.

In the future, I may extend this trait to provide more queryability for elements, e.g. allow an element to provide additional queryable keys and values, so elements can be indexed and queried by more than just type.

Example Implementation

#[derive(Debug)]
struct MyElementType {
    meta_data : HashMap<String, String>
}

impl QWSElement for MyElementType {

    fn meta_data_keys_iter(&self) -> Option<Box<dyn Iterator<Item=QWSMetaData> + '_>> {
        Some(Box::new(self.meta_data.keys().map(|string_key| QWSMetaData::String(string_key.clone()))))
    }

    fn meta_data_values_for_key_iter(&self, key : &QWSMetaData) -> Option<Box<dyn Iterator<Item=QWSMetaData> + '_>> {
        match key {
            QWSMetaData::String(string_key) => {
                self.meta_data.get(string_key).map(|value| { 
                    Box::new(vec![QWSMetaData::String(value.clone())].into_iter()) as Box<dyn Iterator<Item=QWSMetaData> + '_>
                })
            },
            _ => None
        }
    }

    fn as_any(&self) -> &dyn Any {
        self
    }
}

Required methods

fn meta_data_keys_iter(&self) -> Option<Box<dyn Iterator<Item = QWSMetaData>>>

Returns an iterator for the meta-data keys expressed by the element, in order to locate the element through queries.

Warning

Currently any type that implements the Eq and Hash traits can be used as a key. However, we may want to limit it to String objects in the future for compatibility with other database backends.

fn meta_data_values_for_key_iter(
&self,
key: &QWSMetaData
) -> Option<Box<dyn Iterator<Item = QWSMetaData>>>

Returns an iterator for the meta-data values associated with the specified key. This function returns an iterator because a key may have multiple values

fn as_any(&self) -> &dyn Any

Returns the element as a &dyn Any, from which you can use downcast_ref() to get to the original object

Example Usage

// Since we know the element is a QWSElementWrapper<i32>...
let cast_element = found_element.as_any().downcast_ref::<QWSElementWrapper<i32>>().unwrap();

Loading content...
Implementations

impl<'dyn_trait> dyn QWSElement + 'dyn_trait

pub fn get_payload<'a, T: 'static + Debug>(&self) -> Option<&T>

Returns an element's payload, if an element happens to be implemented as a QWSElementWrapper.
Returns None if the type specified doesn't match the payload, or if the element isn't a QWSElementWrapper.

Examples

// Since we know the element payload is an i32
found_element.get_payload::<i32>().unwrap();